whois-api/pyproject.toml
[project]
name = "whois-api"
version = "0.1.0"
description = "FastAPI WHOIS/RDAP API"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "fastapi",
  "httpx",
  "python-whois",
  "uvicorn[standard]",
]

[dependency-groups]
dev = [
  "pytest",
  "pytest-asyncio",
  "ruff",
]

[tool.uv]
default-groups = ["dev"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.pytest.ini_options]
asyncio_mode = "auto"
